+++ frontend/src/components/FileInfoForm.vue
@@ -249,6 +249,9 @@
                                         case 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet':
                                             type = 'Data';
                                             break;
+                                        case 'application/vnd.ms-excel':
+                                            type = 'Data';
+                                            break
                                         case 'text/plain':
                                             type = 'Data';
                                             break;